requirements.txt
fastapi
uvicorn
google-cloud-firestore
firebase_admin
python-dotenv